Fix SelectedItemsList export structure and example imports
