+++ packages/experiments/src/components/SelectedItemsList/SelectedPeopleList/index.ts
@@ -1,0 +1,3 @@
+export * from './SelectedPeopleList';
+export * from './Items/SelectedPersona';
+export * from './Items/SelectedPersona.types';
