+++ packages/experiments/src/components/SelectedItemsList/index.ts
@@ -1,6 +1,5 @@
 export * from './SelectedItemsList.types';
 export * from './SelectedItemsList';
-export * from './SelectedPeopleList/SelectedPeopleList';
-export * from './SelectedPeopleList/Items/SelectedPersona';
-export * from './Items/EditableItem';
-export * from './Items/ItemWithContextMenu';
+export * from './SelectedPeopleList/index';
+export * from './Items/index';
+export * from './utils/copyToClipboard';
